add a timestamp method to c-node-api

+++ apis/c/node/node_api.h
@@ -18,5 +18,5 @@ enum DoraEventType read_dora_event_type(void *dora_event);
 
 void read_dora_input_id(void *dora_event, char **out_ptr, size_t *out_len);
 void read_dora_input_data(void *dora_event, char **out_ptr, size_t *out_len);
-uint64_t read_dora_input_timestamp(void *dora_event);
+unsigned long long read_dora_input_timestamp(void *dora_event);
 int dora_send_output(void *dora_context, char *id_ptr, size_t id_len, char *data_ptr, size_t data_len);
